requirement.txt
python-dotenv==0.19.1
spotipy==2.19.0